# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * 校园表白墙 API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Post {
  id: number;
  content: string;
  nickname: string;
  likeCount: number;
  commentCount: number;
  liked: boolean;
  createdAt: string;
}

export interface PostInput {
  /**
     * @minLength 1
     * @maxLength 500
     */
  content: string;
  /** @maxLength 20 */
  nickname?: string;
}

export interface PostList {
  posts: Post[];
  total: number;
  page: number;
  pageSize: number;
}

export interface Comment {
  id: number;
  postId: number;
  content: string;
  nickname: string;
  createdAt: string;
}

export interface CommentInput {
  /**
     * @minLength 1
     * @maxLength 200
     */
  content: string;
  /** @maxLength 20 */
  nickname?: string;
}

export interface LikeResult {
  liked: boolean;
  likeCount: number;
}

export interface StatsSummary {
  totalPosts: number;
  totalLikes: number;
  totalComments: number;
  todayPosts: number;
}

export type ListPostsParams = {
page?: number;
pageSize?: number;
sort?: ListPostsSort;
};

export type ListPostsSort = typeof ListPostsSort[keyof typeof ListPostsSort];


export const ListPostsSort = {
  latest: 'latest',
  popular: 'popular',
} as const;