# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: 校园表白墙 API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: posts
    description: Confession posts
  - name: likes
    description: Post likes
  - name: comments
    description: Post comments
  - name: stats
    description: Statistics and aggregates
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /posts:
    get:
      operationId: listPosts
      tags: [posts]
      summary: List all confession posts
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
        - name: sort
          in: query
          schema:
            type: string
            enum: [latest, popular]
            default: latest
      responses:
        "200":
          description: List of posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostList"
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a new confession post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"

  /posts/{id}:
    get:
      operationId: getPost
      tags: [posts]
      summary: Get a single post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: The post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "404":
          description: Not found
    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /posts/{id}/like:
    post:
      operationId: toggleLike
      tags: [likes]
      summary: Toggle like on a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated like count
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeResult"

  /posts/{id}/comments:
    get:
      operationId: listComments
      tags: [comments]
      summary: List comments for a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: createComment
      tags: [comments]
      summary: Create a comment on a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Created comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"

  /comments/{commentId}:
    delete:
      operationId: deleteComment
      tags: [comments]
      summary: Delete a comment
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /stats/summary:
    get:
      operationId: getStatsSummary
      tags: [stats]
      summary: Get wall statistics summary
      responses:
        "200":
          description: Summary stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatsSummary"

  /posts/trending:
    get:
      operationId: getTrendingPosts
      tags: [posts]
      summary: Get top trending posts by likes
      responses:
        "200":
          description: Trending posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Post:
      type: object
      required: [id, content, nickname, likeCount, commentCount, liked, createdAt]
      properties:
        id:
          type: integer
        content:
          type: string
        nickname:
          type: string
        likeCount:
          type: integer
        commentCount:
          type: integer
        liked:
          type: boolean
        createdAt:
          type: string
          format: date-time

    PostInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 500
        nickname:
          type: string
          maxLength: 20

    PostList:
      type: object
      required: [posts, total, page, pageSize]
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer

    Comment:
      type: object
      required: [id, postId, content, nickname, createdAt]
      properties:
        id:
          type: integer
        postId:
          type: integer
        content:
          type: string
        nickname:
          type: string
        createdAt:
          type: string
          format: date-time

    CommentInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 200
        nickname:
          type: string
          maxLength: 20

    LikeResult:
      type: object
      required: [liked, likeCount]
      properties:
        liked:
          type: boolean
        likeCount:
          type: integer

    StatsSummary:
      type: object
      required: [totalPosts, totalLikes, totalComments, todayPosts]
      properties:
        totalPosts:
          type: integer
        totalLikes:
          type: integer
        totalComments:
          type: integer
        todayPosts:
          type: integer