# lib/api-zod/src/generated/types/commentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * 校园表白墙 API
 * OpenAPI spec version: 0.1.0
 */

export interface CommentInput {
  /**
     * @minLength 1
     * @maxLength 200
     */
  content: string;
  /** @maxLength 20 */
  nickname?: string;
}