# lib/api-zod/src/generated/types/postInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * 校园表白墙 API
 * OpenAPI spec version: 0.1.0
 */

export interface PostInput {
  /**
     * @minLength 1
     * @maxLength 500
     */
  content: string;
  /** @maxLength 20 */
  nickname?: string;
}