# lib/api-zod/src/generated/types/statsSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * 校园表白墙 API
 * OpenAPI spec version: 0.1.0
 */

export interface StatsSummary {
  totalPosts: number;
  totalLikes: number;
  totalComments: number;
  todayPosts: number;
}